// packages/node-webcodecs/lib/types.ts
export type AllowSharedBufferSource = ArrayBuffer | ArrayBufferView;
// ...
export type PremultiplyAlpha = 'none' | 'premultiply' | 'default';
// ...
export type ColorSpaceConversion = 'default' | 'none';
// ...
export type ImageBufferSource = AllowSharedBufferSource | ReadableStream<Uint8Array>;
// ...
export interface ImageDecoderInit {
  type: string;
  data: ImageBufferSource;
  colorSpaceConversion?: ColorSpaceConversion;
  premultiplyAlpha?: PremultiplyAlpha;
  desiredWidth?: number;
  desiredHeight?: number;
  preferAnimation?: boolean;
  transfer?: ArrayBuffer[];
}

/**
 * WebIDL:
 * dictionary ImageDecodeOptions {
 *   [EnforceRange] unsigned long frameIndex = 0;
 *   boolean completeFramesOnly = true;
 * };
 */
export interface ImageDecodeOptions {
  frameIndex?: number;
  completeFramesOnly?: boolean;
}
// ...
export interface ImageTrack {
  readonly animated: boolean;
  readonly frameCount: number;
  readonly repetitionCount: number; // unrestricted float, Infinity for infinite loop
  selected: boolean;
}

/**
 * WebIDL:
 * interface ImageTrackList {
 *   getter ImageTrack (unsigned long index);
 *   readonly attribute Promise<undefined> ready;
 *   readonly attribute unsigned long length;
 *   readonly attribute long selectedIndex;
 *   readonly attribute ImageTrack? selectedTrack;
 * };
 */
export interface ImageTrackList {
  readonly ready: Promise<void>;
  readonly length: number;
  readonly selectedIndex: number;
  readonly selectedTrack: ImageTrack | null;
  [index: number]: ImageTrack;
}